move out getAuthToken method add static to constants
