+++ src/test/java/CreatingOrderTest.java
@@ -44,15 +44,6 @@ public class CreatingOrderTest extends BaseURIAndAPIs {
                 .post("/api/auth/register");
     }
 
-    @Step("Get auth token")
-    public String getAuthToken(Response response) {
-        return response
-                .then()
-                .extract()
-                .path("accessToken")
-                .toString()
-                .replace("Bearer ", "");
-    }
 
     @Step("Check authorisation user")
     public void checkLoginUser() {
