+++ src/test/java/UserAuthorizationTest.java
@@ -33,17 +33,6 @@ public class UserAuthorizationTest extends BaseURIAndAPIs {
     }
 
 
-
-    @Step("Get auth token")
-    public String getAuthToken(Response response) {
-        return response
-                .then()
-                .extract()
-                .path("accessToken")
-                .toString()
-                .replace("Bearer ", "");
-    }
-
     @Step("Check authorisation user")
     public void checkLoginUser() {
         login(user)
